# Conflicts:
#	Client/src/pages/admin/CoursesManage/CoursesManage.jsx
#	Client/src/pages/admin/CoursesManage/EditCourses.jsx
#	Client/src/pages/admin/CoursesManage/ViewCourses.jsx
#	Client/src/pages/admin/layout-admin/employee-manage/AddEmployeeForm.jsx
#	Client/src/pages/admin/layout-admin/employee-manage/EditEmployeeForm.jsx
#	Client/src/pages/admin/layout-admin/employee-manage/EmployeeLayout.jsx
#	Client/src/pages/admin/layout-admin/employee-manage/EmployeeList.jsx
#	Spring/src/main/java/com/htilssu/sport/controllers/CoachController.java
#	Spring/src/main/java/com/htilssu/sport/services/CoachService.java
#	Spring/src/main/resources/db/migration/V2__.sql
#	Spring/src/main/resources/db/migration/V3__.sql

+++ Spring/src/main/java/com/htilssu/sport/data/dtos/EquipmentTypeDto.java
@@ -1,8 +1,13 @@
 package com.htilssu.sport.data.dtos;
 
 import java.io.Serializable;
+
 /**
  * DTO for {@link com.htilssu.sport.data.models.EquipmentType}
  */
-public record EquipmentTypeDto(Long id,String name,Short amount) implements Serializable {
+public record EquipmentTypeDto(
+        Long id,
+        String name,
+        Integer amount
+) implements Serializable {
 }
